% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectpartitions.R
\name{get_rectangles}
\alias{get_rectangles}
\title{Function that tries to partition the matrix of integers into a minimal set of rectangles.}
\usage{
get_rectangles(m, target_value = 1)
}
\arguments{
\item{m}{Matrix of integer to partition.}

\item{target_value}{Value to use as a marker. Defaults to "1". Any other value than that will be treated as the absense of marker.}
}
\value{
Returns list of rectangles. Each rectangle is a list defined by \code{x0}, \code{x1}, \code{y0} and \code{y1}.
}
\description{
Function that tries to partition the matrix of integers into a minimal set of rectangles.
}
