% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectpartitions.R
\name{get_rectangles_shuffle}
\alias{get_rectangles_shuffle}
\title{Function that re-shuffles columns and rows of the integer matrix to find the best partitionioning into a minimal set of rectangles.}
\usage{
get_rectangles_shuffle(m, target_value = 1, method = "single")
}
\arguments{
\item{m}{Matrix of integer to partition.}

\item{target_value}{Value to use as a marker. Defaults to "1". Any other value than that will be treated as the absense of marker.}

\item{method}{Shuffling is done by means of hierarchical clustering. This parameter denotes the clustering algorithm to use. Defaults to \code{'single'}.}
}
\value{
Returns list of rectangles. Each rectangle is a list defined by \code{x0}, \code{x1}, \code{y0} and \code{y1}.
}
\description{
Function that re-shuffles columns and rows of the integer matrix to find the best partitionioning into a minimal set of rectangles.
}
